get column width on windows
