+++ textcol.go
@@ -5,8 +5,6 @@ import (
 	"io"
 	"math"
 	"os"
-	"os/exec"
-	"strconv"
 	"strings"
 	"unicode/utf8"
 
@@ -73,17 +71,6 @@ func PrintColumns(strs *[]string, margin int) {
 			fmt.Fprintf(Output, marginStr)
 		}
 	}
-}
-
-func getTermWidth() int {
-	cmd := exec.Command("stty", "size")
-	cmd.Stdin = os.Stdin
-	out, err1 := cmd.Output()
-	check(err1)
-	numsStr := strings.Trim(string(out), "\n ")
-	width, err2 := strconv.Atoi(strings.Split(numsStr, " ")[1])
-	check(err2)
-	return width
 }
 
 func calculateTableSize(width, margin, maxLength, numCells int) (int, int) {
